backend/requirements/base.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
pydantic-settings==2.5.2
sqlalchemy==2.0.35
alembic==1.13.3
asyncpg==0.29.0
psycopg[binary]==3.2.3
psycopg2-binary==2.9.10
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2  
python-multipart==0.0.12
aiofiles==24.1.0
httpx==0.27.2
celery==5.4.0
redis==5.2.0
ezkl==10.2.9
web3==7.4.0
python-dotenv==1.0.1
python-magic==0.4.27
aiofiles==24.1.0
onnx==1.17.0
itsdangerous>=2.1.2,<3.0
psutil==7.1.0
email-validator==2.3.0